infra/requirements.txt
aws-cdk-lib==2.120.0
constructs>=10.0.0,<11.0.0
pyjwt[crypto]>=2.8.0,<3.0.0
